scripts/ingest/requirements.txt
openai>=1.40.0
supabase>=2.5.1
pymupdf>=1.24.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
tqdm>=4.66.0
